updates per comments
